requirements.txt
pytz==2019.3
tzlocal==5.0.1
pypandoc==1.4
feedparser==5.2.1
pillow==6.2.1
morss